requirements.txt
discord.py
aiohttp
aiohttp-proxy
asyncpg
fastapi
